requirements.txt
mesa == 3.0.3
matplotlib == 3.5.1
numpy == 1.21.5
networkx == 3.3
protobuf == 3.12.4
torch == 2.5.0
pyyaml == 6.0
